Remove pointer to Process in Connections class
In rust it's not trivial to keep a reference from the Connections class
to the Process class which has Connections as a member. Therefore, the
methods from Connections should be moved into Process and this pointer
removed. This commit does that. The Connections class is still not
removed for future reference.